refactor: abstracted request generation logic
Abstracted request generation logic to github handler as it manager the
username and token value for the user. It also manager the repository
fetch and pulls on the repository as the entity is done by the user in
the real world and thus it should have the data to perform those tasks.