fix: bug in name of PR created by dependabot
dependabot created some pr with the name in pattern `bump` (lowercase
bump) and that lead to failure in detecting the PRs. Also there were
cases of change in PR title.

Refactored the logic to use the lowercase comparison and not look for
start but rather check for the words in the string in the order.